+++ config/config.go
@@ -27,10 +27,18 @@ var (
 
 // Config defines the configuration for the various elements of collector or agent.
 type Config struct {
-	Receivers
-	Exporters
-	Processors
-	Extensions
+	// Receivers is a map of ComponentID to Receivers.
+	Receivers map[ComponentID]Receiver
+
+	// Exporters is a map of ComponentID to Exporters.
+	Exporters map[ComponentID]Exporter
+
+	// Processors is a map of ComponentID to Processors.
+	Processors map[ComponentID]Processor
+
+	// Extensions is a map of ComponentID to extensions.
+	Extensions map[ComponentID]Extension
+
 	Service
 }
 
@@ -42,43 +50,43 @@ var _ validatable = (*Config)(nil)
 // invalid cases that we currently don't check for but which we may want to add in
 // the future (e.g. disallowing receiving and exporting on the same endpoint).
 func (cfg *Config) Validate() error {
-	// Currently there is no default receiver enabled.
+	// Currently, there is no default receiver enabled.
 	// The configuration must specify at least one receiver to be valid.
 	if len(cfg.Receivers) == 0 {
 		return errMissingReceivers
 	}
 
 	// Validate the receiver configuration.
-	for recv, recvCfg := range cfg.Receivers {
+	for recvID, recvCfg := range cfg.Receivers {
 		if err := recvCfg.Validate(); err != nil {
-			return fmt.Errorf("receiver \"%s\" has invalid configuration: %w", recv, err)
+			return fmt.Errorf("receiver %q has invalid configuration: %w", recvID, err)
 		}
 	}
 
-	// Currently there is no default exporter enabled.
+	// Currently, there is no default exporter enabled.
 	// The configuration must specify at least one exporter to be valid.
 	if len(cfg.Exporters) == 0 {
 		return errMissingExporters
 	}
 
 	// Validate the exporter configuration.
-	for exp, expCfg := range cfg.Exporters {
+	for expID, expCfg := range cfg.Exporters {
 		if err := expCfg.Validate(); err != nil {
-			return fmt.Errorf("exporter \"%s\" has invalid configuration: %w", exp, err)
+			return fmt.Errorf("exporter %q has invalid configuration: %w", expID, err)
 		}
 	}
 
 	// Validate the processor configuration.
-	for proc, procCfg := range cfg.Processors {
+	for procID, procCfg := range cfg.Processors {
 		if err := procCfg.Validate(); err != nil {
-			return fmt.Errorf("processor \"%s\" has invalid configuration: %w", proc, err)
+			return fmt.Errorf("processor %q has invalid configuration: %w", procID, err)
 		}
 	}
 
 	// Validate the extension configuration.
-	for ext, extCfg := range cfg.Extensions {
+	for extID, extCfg := range cfg.Extensions {
 		if err := extCfg.Validate(); err != nil {
-			return fmt.Errorf("extension \"%s\" has invalid configuration: %w", ext, err)
+			return fmt.Errorf("extension %q has invalid configuration: %w", extID, err)
 		}
 	}